% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.create.tree.R
\name{batch.create.tree}
\alias{batch.create.tree}
\title{Create tree structures for group search solutions.}
\usage{
batch.create.tree(hist, ind_hist, ind_fit, subgroup, names.ts_list, sub)
}
\description{
Create tree structures for group search solutions.
}
\keyword{internal}
