% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmented.variables.R
\name{constrainAugmentedSyntax}
\alias{constrainAugmentedSyntax}
\title{Fix prohibited augmented-variable paths to zero for one individual}
\usage{
constrainAugmentedSyntax(syntax, augmented_vars)
}
\arguments{
\item{syntax}{Lavaan model syntax.}

\item{augmented_vars}{Base names of augmented variables for one individual.}
}
\description{
Fix prohibited augmented-variable paths to zero for one individual
}
\keyword{internal}
