% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupConvolve.R
\name{convolve}
\alias{convolve}
\alias{convolveFIR}
\alias{setupConvolve}
\title{Group iterative multiple model estimation.}
\usage{
convolveFIR(ts_list = NULL, 
     varLabels = NULL, 
     conv_length = 16, 
     conv_interval = 1)
}
\arguments{
\item{ts_list}{a list of dataframes.}

\item{varLabels}{a list of variable sets. Contains varLabels$coln, all column names, varLabels$conv, 
the names of variables to convolve, and varLabels$exog, a list of exogenous variables (if any).}

\item{conv_length}{Expected response length in seconds. For functional MRI BOLD, 16 seconds (default) is typical
for the hemodynamic response function.}

\item{conv_interval}{Interval between data acquisition. Currently must be a constant. For 
fMRI studies, this is the repetition time. Defaults to 1.}
}
\description{
This function estimates the basis vectors related to responses following 
a binary impulse and convolves that binary impulse vector.
}
\keyword{setupConvolve}
