% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.tree.R
\name{create.tree}
\alias{create.tree}
\title{Create structure of group search solutions.}
\usage{
create.tree(history, subgroup, individual = FALSE, all.ind = FALSE)
}
\description{
Create structure of group search solutions.
}
\keyword{internal}
