% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.grid.unique.R
\name{expand.grid.unique}
\alias{expand.grid.unique}
\title{Provides unique combinations of two vectors.}
\usage{
expand.grid.unique(x, y, incl.eq = TRUE)
}
\arguments{
\item{x}{A character vector containing variable names.}

\item{y}{A character vector containing variable names.}

\item{incl.eq}{Logical. TRUE means that combinations are kept where
a variable appears twice.}
}
\value{
The unique combinations of the variable names. Used in syntax
creation.
}
\description{
Provides unique combinations of two vectors.
}
\keyword{internal}
