% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.params.R
\name{get.params}
\alias{get.params}
\title{Grabs final coefficients for each individual.}
\usage{
get.params(dat, grp, ind, k)
}
\arguments{
\item{dat}{A list containing information created in setup().}

\item{grp}{A list containing group-level information. NULL in aggSEM and
indSEM.}

\item{ind}{A list containing individual- and (potentially) subgroup-level
information.}

\item{k}{The counter indicating the individual.}
}
\value{
Individual-level information on fit, coefficients, and plots.
}
\description{
Grabs final coefficients for each individual.
}
\keyword{internal}
