% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gimmeInteractive.R
\name{gimmeInteractive}
\alias{gimmeInteractive}
\title{Graphical User Interface designed for use with the gimme package}
\usage{
gimmeInteractive()
}
\description{
This function launches a graphical user interface which calls the main functions
for the gimme package. gimmeInteractive() requires the package gWidgetsRGtk2, which requires the GTK+ library.
Please install this package before using gimmeInteractive().
}
\author{
Hallie Pike
}

