\name{het4}
\alias{het4}
\docType{data}
\title{
In-house simulated time-series data with heterogeneity.
}
\description{
This data contains simulated time-series data for a single individual with 200 time points and 10 variables, or regions of interest.
}
\usage{data("het4")}
\format{
  A data frame with 200 observations on 10 variables.
}
\keyword{datasets}
