% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimme-data.R
\docType{data}
\name{HRFsim}
\alias{HRFsim}
\title{Convolution with sFIR example, direct and modulatory effect}
\format{A list of data frames with 25 individuals, who each have 500 observations on 5 variables.}
\usage{
HRFsim
}
\description{
This object contains a list of simulated time series data designed to be used for HRF-GIMME 
and has twenty-five individuals with 500 time points and 4 variables, or brain regions of interest, and
one task vector i.e, exogenous variable.
}
\keyword{datasets}
