% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmented.variables.R
\name{isAugmentedPath}
\alias{isAugmentedPath}
\title{Identify relations prohibited for an augmented variable}
\usage{
isAugmentedPath(lhs, op, rhs, augmented_vars)
}
\arguments{
\item{lhs}{Left-hand-side variable names.}

\item{op}{Lavaan operators.}

\item{rhs}{Right-hand-side variable names.}

\item{augmented_vars}{Base names of augmented variables for one individual.}
}
\description{
Identify relations prohibited for an augmented variable
}
\keyword{internal}
