% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmented.variables.R
\name{maskAugmentedMIs}
\alias{maskAugmentedMIs}
\title{Remove MI/EPC evidence for prohibited augmented-variable paths}
\usage{
maskAugmentedMIs(mis, augmented_vars)
}
\arguments{
\item{mis}{A modification-index data frame.}

\item{augmented_vars}{Base names of augmented variables for one individual.}
}
\description{
Remove MI/EPC evidence for prohibited augmented-variable paths
}
\keyword{internal}
