% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmented.variables.R
\name{maskAugmentedZs}
\alias{maskAugmentedZs}
\title{Remove significance evidence for prohibited augmented-variable paths}
\usage{
maskAugmentedZs(zs, augmented_vars)
}
\arguments{
\item{zs}{A standardized-solution data frame.}

\item{augmented_vars}{Base names of augmented variables for one individual.}
}
\description{
Remove significance evidence for prohibited augmented-variable paths
}
\keyword{internal}
