% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimme-data.R
\docType{data}
\name{ms.fit}
\alias{ms.fit}
\title{Fitted gimme object with multiple solutions}
\format{A fitted gimme object, where multiple solutions gimme was used.}
\usage{
ms.fit
}
\description{
This object contains a fitted gimme object where multiple solutions gimme was used.  The simulated
data had 25 individuals, each with 100 time points.
}
\keyword{datasets}
