% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.gimme.R
\name{predict.gimme}
\alias{predict.gimme}
\title{GIMME Predicted Values.}
\usage{
predict.gimme(x)
}
\arguments{
\item{x}{A fitted gimme object.}
}
\value{
List of data frames. Each data frame contains the predicted values of
  a subject in the data.
}
\description{
This function calculates the predicted values of a fitted gimme
  model.
}
\examples{
 \dontrun{
paths <- 'V2 ~ V1
          V3 ~ V4lag'

fit <- gimmeSEM(data     = simData,
                out      = "C:/simData_out",
                subgroup = TRUE, 
                paths    = paths)

predictions <- predict.gimme(fit)
 }
}
\author{
Sebastian Castro-Alvarez
}
\keyword{internal}
