% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return.mis.R
\name{return.mis}
\alias{return.mis}
\title{Returns MIs from lavaan fit object.}
\usage{
return.mis(fit, elig_paths, augmented_vars = character())
}
\arguments{
\item{fit}{An object from lavaan.}

\item{elig_paths}{Paths eligible at the current search stage.}

\item{augmented_vars}{Base names of augmented variables for this individual.}
}
\value{
If successful, returns MIs for an individual. If unsuccessful, 
returns NA.
}
\description{
Returns MIs from lavaan fit object.
}
\keyword{internal}
