% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmented.variables.R
\name{setupAugmentedData}
\alias{setupAugmentedData}
\title{Prepare data containing augmented variables}
\usage{
setupAugmentedData(data)
}
\arguments{
\item{data}{A list of individual data sets with augmented columns ending
in "_AUG".}
}
\description{
Prepare data containing augmented variables
}
\keyword{internal}
