% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupConvolve.R
\name{setupConvolve}
\alias{setupConvolve}
\title{Convolve data.}
\usage{
setupConvolve(ts_list = NULL, varLabels = NULL, conv_length = 16,
  conv_interval = 1)
}
\arguments{
\item{ts_list}{a list of dataframes.}

\item{conv_length}{Expected response length in seconds. For functional MRI BOLD, 16 seconds (default) is typical
for the hemodynamic response function.}

\item{conv_interval}{Interval between data acquisition. Currently must be a constant. For 
fMRI studies, this is the repetition time. Defaults to 1.}

\item{varLebels}{a list of variable sets.}
}
\description{
Convolve data.
}
\keyword{internal}
