% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupDataLists.R
\name{setupDataLists}
\alias{setupDataLists}
\title{Create a list of dataframes}
\usage{
setupDataLists(data, ctrlOpts = NULL, lv_model = NULL)
}
\arguments{
\item{data}{a list or directory.}

\item{ctrlOpts}{a lit of control options.}
}
\description{
Create a list of dataframes
}
\keyword{internal}
