% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupFinalDataChecks.R
\name{setupFinalDataChecks}
\alias{setupFinalDataChecks}
\title{Do some final checks on the data.}
\usage{
setupFinalDataChecks(data)
}
\description{
Do some final checks on the data.
}
\keyword{internal}
