% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupPrelimDataChecks.R
\name{setupPrelimDataChecks}
\alias{setupPrelimDataChecks}
\title{Do some preliminary checks on the data.}
\usage{
setupPrelimDataChecks(df)
}
\description{
Do some preliminary checks on the data.
}
\keyword{internal}
