% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupPrepPaths.R
\name{setupPrepPaths}
\alias{setupPrepPaths}
\title{Allows user to open and close certain paths.}
\usage{
setupPrepPaths(paths, varLabels, ctrlOpts)
}
\arguments{
\item{paths}{\code{lavaan}-style syntax containing paths with which
to begin model estimation (optional). That is, Y~X indicates that Y
is regressed on X, or X predicts Y. If no header is used,
then variables should be referred to with V followed (with no separation)
by the column number. If a header is used, variables should be referred to using 
variable names. To reference lag variables, "lag" should be added to the 
end of the variable name with no separation. Defaults to NULL.}
}
\description{
Allows user to open and close certain paths.
}
\keyword{internal}
