% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupTransformData.R
\name{setupTransformData}
\alias{setupTransformData}
\title{Transform raw data as required.}
\usage{
setupTransformData(ts_list = NULL, varLabels = NULL, ctrlOpts = NULL)
}
\arguments{
\item{ts_list}{a list or directory}

\item{varLabels}{Variable labels.}

\item{ctrlOpts}{List used in setup function.}
}
\description{
Transform raw data as required.
}
\keyword{internal}
