% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sFIR.R
\name{sFIR}
\alias{sFIR}
\title{Estimate response function for each person using smoothed Finite Impulse Response.}
\usage{
sFIR(data, stimuli, response_length = 16, interval = 1)
}
\arguments{
\item{data}{The data to be used to estimate response function}

\item{stimuli}{A vector containing '0' when the stimuli of interest is not present 
and '1' otherwise. Number of observations across time must equal the length of data.}

\item{interval}{Time between observations; for fMRI this is the repetition time. 
Defaults to 1.}
}
\value{
Shape of response function and convolved time series vector.
}
\description{
Estimate response function for each person using smoothed Finite Impulse Response.
}
\keyword{internal}
