% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimme-data.R
\docType{data}
\name{simData}
\alias{simData}
\title{Large example, heterogeneous data, group, subgroup, and individual level effects.}
\format{A list of data frames with 25 individuals, who each have 200 observations on 10 variables.}
\usage{
simData
}
\description{
This object contains a list of simulated time series data for twenty-five individuals 
with 200 time points and 10 variables, or regions of interest.
}
\keyword{datasets}
