% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimme-data.R
\docType{data}
\name{simDataLV}
\alias{simDataLV}
\title{Latent variable example, heterogeneous data, group, subgroup level effects.}
\format{A list of data frames with 20 individuals, who each have 500 observations on 9 variables.}
\usage{
simDataLV
}
\description{
This object contains a list of simulated time series data for twenty individuals 
with 500 time points and 9 variables, or regions of interest.
}
\keyword{datasets}
