% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPathsCounts.R
\name{summaryPathsCounts}
\alias{summaryPathsCounts}
\title{Create summary matrix of path counts and subgroup plots}
\usage{
summaryPathsCounts(dat, grp, store, sub, sub_spec)
}
\arguments{
\item{dat}{A list containing information created in setup().}

\item{grp}{A list containing group-level information. NULL in aggSEM and
indSEM.}

\item{store}{A list containing output from indiv.search().}

\item{sub}{A list containing subgroup information.}

\item{sub_spec}{A list containing information specific to each subgroup.}
}
\value{
Aggregated information, such as counts, levels, and plots.
}
\description{
Create summary matrix of path counts and subgroup plots
}
\keyword{internal}
