% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimme-data.R
\docType{data}
\name{ts}
\alias{ts}
\title{Small example, heterogeneous data, group and individual level effects}
\format{A list of data frames with 5 individuals, who each have 50 observations on 3 variables.}
\usage{
ts
}
\description{
This object contains a list of simulated time series data for five individuals 
with 50 time points and 3 variables, or regions of interest.
}
\keyword{datasets}
