% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ts3}
\alias{ts3}
\title{ts3}
\format{A data frame with 50 observations on 3 variables.}
\usage{
ts3
}
\description{
This data contains simulated time-series data for a single individual with 50 time points and 3 variables, or regions of interest.
}
\keyword{datasets}

