% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2e.R
\name{w2e}
\alias{w2e}
\title{Create edge list from weight matrix.}
\usage{
w2e(x)
}
\arguments{
\item{x}{The coefficient matrix from an individual}
}
\value{
A list of all non-zero edges to feed to qgraph
}
\description{
Create edge list from weight matrix.
}
\keyword{internal}
