% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggSEM.R
\name{aggSEM}
\alias{aggSEM}
\title{Group-level structural equation model search.}
\usage{
aggSEM(data   = "",
       out    = "",
       sep    = "",
       header = ,
       ar     = TRUE,
       plot   = TRUE,
       paths  = NULL,
       exogenous = NULL, 
       conv_vars        = NULL,
       conv_length      = 16, 
       conv_interval    = 1, 
       mult_vars        = NULL,
       mean_center_mult = FALSE,
       standardize      = FALSE,
       hybrid = FALSE)
}
\arguments{
\item{data}{The path to the directory where the data files are located, 
or the name of the list containing each individual's time series. 
Each file or matrix must contain one matrix 
for each individual containing a T (time) by 
p (number of variables) matrix where the columns represent variables and
the rows represent time. If in list form, each item in the list (i.e., matrix) must be named.}

\item{out}{The path to the directory where the results will 
be stored (optional). If specified, a copy of output files will be replaced
in directory. If directory at specified path does not
exist, it will be created.}

\item{sep}{The spacing of the data files when data are in a directory. "" indicates space-delimited, 
"/t" indicates tab-delimited, "," indicates comma delimited. 
Only necessary to specify if reading data in from physical directory.}

\item{header}{Logical. Indicate TRUE for data files with a header. 
Only necessary to specify if reading data in from physical directory.}

\item{ar}{Logical. If TRUE, begins search for group model with 
autoregressive (AR) paths open. Defaults to TRUE.}

\item{plot}{Logical. If TRUE, figures depicting relations among variables 
of interest will automatically be created. For aggregate-level plot, red 
paths represent positive weights and blue paths represent negative weights. 
Dashed lines denote lagged relations (lag 1) and solid lines are contemporaneous 
(lag 0). Defaults to TRUE.}

\item{paths}{\code{lavaan}-style syntax containing paths with which to begin
model estimation (optional). That is, Y~X indicates that Y is regressed on X, or X
predicts Y. If no header is used, then variables should be referred to
with V followed (with no separation) by the column number. If a
header is used, variables should be referred to using variable names. 
To reference lag variables, "lag" should be added to the end of the variable 
name with no separation. Defaults to NULL.}

\item{exogenous}{Vector of variable names to be treated as exogenous (optional).
That is, exogenous variable X can predict Y  but cannot be predicted by Y.  
If no header is used, then variables should be referred to with V followed 
(with no separation) by the column number. If a header is used, variables 
should be referred to using variable names.  Defaults to NULL.}

\item{conv_vars}{Vector of variable names to be convolved via smoothed Finite Impulse 
Response (sFIR). Defaults to NULL.}

\item{conv_length}{Expected response length in seconds. For functional MRI BOLD, 16 seconds (default) is typical
for the hemodynamic response function.}

\item{conv_interval}{Interval between data acquisition. Currently must be a constant. For 
fMRI studies, this is the repetition time. Defaults to 1.}

\item{mult_vars}{Vector of variable names to be multiplied to explore bilinear/modulatory
effects (optional). All multiplied variables will be treated as exogenous (X can predict
Y but cannot be predicted by Y). Within the vector, multiplication of two variables should be
indicated with an asterik (e.g. V1*V2). If no header is used, variables should be referred to with 
V followed by the column number (with no separation). If a header is used, each variable should be
referred to using variable names. If multiplication with the lag 1 of a variable is desired, the 
variable name should be followed by "lag" with no separation (e.g. V1*V2lag). Note that if
multiplied variables are desired, at least one variable in the dataset must be specified as exogenous.
Defaults to NULL.}

\item{mean_center_mult}{Logical. If TRUE, the variables indicated in mult_vars will be mean-centered
before being multiplied together. Defaults to FALSE.}

\item{standardize}{Logical. If TRUE, all variables will be standardized to have a mean of zero and a
standard deviation of one. Defaults to FALSE.}

\item{hybrid}{Logical. If TRUE, enables hybrid-VAR models where both directed contemporaneous paths and contemporaneous
covariances among residuals are candidate relations in the search space. Defaults to FALSE.
 @details
 In main output directory:
 \itemize{
 \item{\strong{allBetas}} Matrix. Contains estimates for each path in the
 aggregate-level model. The row variable is the outcome and the column 
 variable is the predictor variable.
 \item{\strong{allStdErrors}} Matrix. Contains standard errors for each path 
 in the aggregate-level model. The row variable is the outcome and the column
 variable is the predictor variable.
 \item{\strong{allPathEstimates}} {Contains estimate, standard error,
 p-value, and z-value for each path for the concatenated data.}
 \item{\strong{summaryFit}} {Contains model fit information for the 
 aggregate-level model.}
 \item{\strong{summaryPathsPlot}} Contains aggregate-level plot. Red paths 
 represent positive weights and blue paths represent negative weights.
}}
}
\description{
Concatenates all individual-level data files and 
fits a group model to the data.
}
\examples{
\dontrun{
exFit <- aggSEM(data = ts)
}
\dontshow{
load(system.file("extdata", "sysdata.rda", package = "gimme"))
}
plot(exFit)
}
\author{
Stephanie Lane
}
