% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_funcs.R
\name{final.org}
\alias{final.org}
\title{Wrapup, create output files.}
\usage{
final.org(dat, grp, ind, sub, sub_spec, store)
}
\arguments{
\item{dat}{A list containing information created in setup().}

\item{grp}{A list containing group-level information. NULL in aggSEM and
indSEM.}

\item{ind}{A list containing individual- and (potentially) subgroup-level
information.}

\item{sub}{A list containing subgroup information.}

\item{sub_spec}{A list containing information specific to each subgroup.}

\item{store}{A list containing output from indiv.search().}
}
\value{
Aggregated information, such as counts, levels, and plots.
}
\description{
Wrapup, create output files.
}
\keyword{internal}
