% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.model.R
\name{fit.model}
\alias{fit.model}
\title{Attempt to fit lavaan model.}
\usage{
fit.model(syntax, data_file)
}
\arguments{
\item{syntax}{A character vector containing syntax.}

\item{data_file}{A data frame containing individual data set.}
}
\value{
If successful, returns fitted lavaan object. If not successful,
catches and returns error message.
}
\description{
Attempt to fit lavaan model.
}
\keyword{internal}
