\name{gimme-package}
\alias{gimme-package}
\alias{gimme-package}
\docType{package}
\title{
gimme: group iterative multiple model estimation
}
\description{
This package contains functions to identify group- and individual-level structural equation models. 
}
\details{
\tabular{ll}{
Package: \tab gimme\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-15\cr
License: \tab GPL-2 \cr
}
The major functions within the gimme package all require the user to specify the directory containing the data and a directory for output to be stored. 
}
\author{
Stephanie Lane

Stephanie Lane  <slane@unc.edu>
}
\references{
Gates & Molenaar, 2012
}

\keyword{ package }

