% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimmeInteractive.R
\name{gimmeInteractive}
\alias{gimmeInteractive}
\alias{gimmeGUI}
\title{Graphical User Interface designed for use with the gimme package}
\usage{
gimmeInteractive()
}
\description{
This function launches a graphical user interface which calls the main functions
for the gimme package. gimmeInteractive() requires the package gWidgetsRGtk2, which requires the GTK+ library. 
Please install this package before using gimmeInteractive().
}
\details{
Returns a simple text file, guiSyntax.txt, containing a copy of the syntax submitted by the GUI. Placed in output folder specified by the user.
}
\author{
Hallie Pike, Stephanie Lane
}
