% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiREG-data.R
\docType{data}
\name{HRFsim}
\alias{HRFsim}
\title{Small Node example with exogenous variable}
\format{A list of data frames with 25 individuals, who each have 500 observations on 5 variables.}
\usage{
HRFsim
}
\description{
This object contains a list of time series data for twenty-five individuals 
with 500 time points and 5 variables.
}
\keyword{datasets}
