% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimme-data.R
\docType{data}
\name{HRFsim}
\alias{HRFsim}
\title{Hemodynamic Response Function (HRF) GIMME example.}
\format{
A list of data frames with 25 individuals, who each have 500 observations on 5 variables.
}
\usage{
HRFsim
}
\description{
This object contains a list of simulated time series data for twenty-five individuals.
Each data set has 500 time points and five variables. The fifth variable represents an onset 
vector for stimulation.
}
\keyword{datasets}
