\name{indSEM}
\alias{indSEM}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Individual-level structural equation model search.
}
\description{
This function identifies structural equation models for each individual. It does not utilize any shared information from the group.
}
\usage{
indSEM(data, sep, header, out, ar, plot)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
Path to the directory where the data files are located. Each file must contain one matrix for each individual containing a T (time) by p (number of variables) matrix where the columns represent variables and the rows represent time.
}
  \item{sep}{
The spacing of the data files. "" indicates space-delimited, "/t" indicates tab-delimited, "," indicates comma delimited.
}
  \item{header}{
Logical. Indicate TRUE for data files with a header.
}
  \item{out}{
The path to the directory where the results will be stored. This directory must be generated by the user prior to running the function.
}
  \item{ar}{
Logical. If TRUE, begins search for individual models with autoregressive (AR) paths open. Defaults to FALSE.
}
  \item{plot}{
Logical. If TRUE, graphs depicting relations among variables of interest will automatically be created. Defaults to TRUE.
}
}
\value{
\item{all.elements.summary}{Contains summary information for paths identified at the individual level.}
\item{all.elements}{Contains information for all paths identified at the individual level.}
\item{all.fit}{Contains model fit information for individual-level models.}
}
\author{Stephanie Lane}
\examples{
\dontrun{
indSEM.out <- indSEM(data = "C:/data100",
              sep = ",",
              header = FALSE,
              out = "C:/data100_indSEM_out",
              ar = TRUE,
              plot = TRUE)
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{indSEM}

