% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indSEM.R
\name{indSEM}
\alias{indSEM}
\title{Individual-level structural equation model search.}
\usage{
indSEM(data   = "",
       out    = "",
       sep    = "",
       header = ,
       ar     = TRUE,
       plot   = TRUE,
       paths  = NULL)
}
\arguments{
\item{data}{The path to the directory where the data files are located. Each file must contain one matrix for
each individual containing a T (time) by p (number of variables) matrix where the columns represent
variables and the rows represent time.}

\item{out}{The path to the directory where the results will be stored. This directory must be generated by
the user prior to running the function.}

\item{sep}{The spacing of the data files. "" indicates space-delimited, "/t" indicates tab-delimited, ","
indicates comma delimited.}

\item{header}{Logical. Indicate TRUE for data files with a header.}

\item{ar}{Logical. If TRUE, begins search for individual models with autoregressive (AR) paths open.
Defaults to TRUE.}

\item{plot}{Logical. If TRUE, graphs depicting relations among variables of interest will automatically be
created. Defaults to TRUE. For individual-level plots, red paths represent positive weights and blue paths represent negative weights.}

\item{paths}{lavaan-style syntax containing paths with which to begin model estimation. That is, Y~X indicates that Y is regressed on X, or X predicts Y. If no header is used,
then variables should be referred to with V followed (with no separation) by the column number. If a
header is used, variables should be referred to using variable names. To reference lag variables, "lag"
should be added to the end of the variable name with no separation. Defaults to NULL.}
}
\description{
This function identifies structural equation models for each individual. It does not utilize
any shared information from the sample.
}
\details{
In main output directory:
 \itemize{
 \item{\strong{indivPathEstimates}} {Contains estimate, standard error, p-value, and z-value for each path for each individual}
 \item{\strong{summaryFit}} {Contains model fit information for individual-level models. }
 \item{\strong{summaryPathCountMatrix}} Contains counts of total number of paths, both contemporaneous and lagged, estimated for the sample. The row variable is the outcome and the column variable is the predictor variable.
 \item{\strong{summaryPathCounts}} {Contains summary count information for paths identified at the individual-level.}
 \item{\strong{summaryPathsPlot}} Contains counts of total number of paths, both contemporaneous and lagged, estimated for the sample. The row variable is the outcome and the column variable is the predictor variable.
 }
 In individual output directory (where \strong{\emph{id}} represents the original file name for each individual):
 \itemize{
 \item{\strong{\emph{id}Betas}} Contains individual-level estimates of each path for each individual.
 \item{\strong{\emph{id}StdErrors}} Contains individual-level standard errors for each path for each individual.
 \item{\strong{\emph{id}Plot}} Contains individual-level plots. Red paths represent positive weights and blue paths represent negative weights.
}
}
\examples{
 \dontrun{
indSEM(data   = "C:/data100",
       out    = "C:/data100_indSEM_out",
       sep    = ",",
       header = FALSE,
       ar     = TRUE,
       plot   = TRUE,
       paths  = NULL)
 }
}
\author{
Stephanie Lane
}
\keyword{indSEM}

