% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_funcs.R
\name{return.mis}
\alias{return.mis}
\title{Returns MIs from lavaan fit object.}
\usage{
return.mis(fit)
}
\arguments{
\item{fit}{An object from lavaan.}
}
\value{
If successful, returns MIs for an individual. If unsuccessful, 
returns NA.
}
\description{
Returns MIs from lavaan fit object.
}
\keyword{internal}
