% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return.zs.R
\name{return.zs}
\alias{return.zs}
\title{Returns z values from lavaan fit object.}
\usage{
return.zs(fit)
}
\arguments{
\item{fit}{An object from lavaan.}
}
\value{
If successful, returns z values for an individual. If unsuccessful, 
returns NA.
}
\description{
Returns z values from lavaan fit object.
}
\keyword{internal}
