% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroupStage.R
\name{subgroupStage}
\alias{subgroupStage}
\title{Create structure of group search solutions.}
\usage{
subgroupStage(dat, grp, confirm_subgroup, elig_paths, sub_feature,
  sub_method, ms_tol, ms_allow)
}
\description{
Create structure of group search solutions.
}
\keyword{internal}
