% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimme-data.R
\docType{data}
\name{ts2}
\alias{ts2}
\title{Example heterogeneous data, 2}
\format{A data frame with 50 observations on 3 variables.}
\usage{
ts2
}
\description{
This data contains simulated time-series data for a single individual with 50 time points and 3 variables, or regions of interest.
}
\keyword{datasets}

